Introduce AbstractContractCallServiceTest
